Reorganize the documentation in order to have a "book"
The reorganization follows the ideas expressed in the issue
https://github.com/ra3xdh/qucs_s_docs/issues/3

You can compile the full documentation using

    latexmk main-en.tex

in the top directory.

There are a lot of repeated things, and the part that I moved to
appendix (called "original documentations") misses several images
and should be adjusted. My idea is that, step by step, the contents
should be moved in normal chapters following the "tutorial" one.

I reduced the depth of the structure to three; still there are lots
of warnings, too much direct "visual" typesetting, repeated anchors,
and the bibliography is not working.

I have not touched the russian part --- I suppose that people that
understand the language can add a `main-ru.tex` and organize it in a
similar way; the same for possible futre translations to other
languages.